+++ Changelog.md
@@ -21,6 +21,8 @@ Breaking Changes:
 * Remove special `:if`/`:unless` filtering metadata. (Phil Pirozhkov, #2870)
 * Remove deprecated `color` configuration option and `--color` command line
   option. (Phil Pirozhkov, #2864)
+* Remove `it_should_behave_like` nested shared group method and
+  `alias_it_should_behave_like_to` configuration option. (Phil Pirozhkov, #2864)
 
 Enhancements: